Consolidate dockerignore generation.
